api-consumer/pyproject.toml
[project]
name = "traffic-visualisation"
version = "0.1.0"
description = "API of data-comsumer -- recieves the json data packages from producer and sends it to visualisation 'server'"
requires-python = ">=3.11"
dependencies = [
    "eventlet>=0.39.1",
    "flask==3.1.0",
    "flask-cors>=5.0.1",
    "gunicorn>=23.0.0",
    "pydantic>=2.11.1",
    "python-dotenv==1.0.1",
    "requests>=2.32.3",
]
